requirements.txt
passlib==1.7.1
Flask==0.12.2
Flask_Login==0.4.0
Flask_SQLAlchemy==2.2
sqlalchemy_migrate==0.11.0
pytest==3.1.3
psycopg2==2.7.1
flask-cors==3.0.3
gevent==1.2.2
flask-cache==0.13.1
flask-mail==0.9.1
redis
bcrypt
flask_paginate
flask_bcrypt
flask_bootstrap